rename to SqlLightningStore operations with SQLAlchemy. - Refactored existing tests to utilize the new SqlLightningStore, ensuring compatibility with the new implementation. - Adjusted timeout behavior in tests to align with the new polling intervals.
